+++ ESFramework/UIKit/Controller/ESStoreProductViewControllerManager.m
@@ -17,7 +17,15 @@
 
 @implementation ESStoreProductViewControllerManager
 
-ES_SINGLETON_IMP(sharedManager);
++ (instancetype)sharedManager
+{
+    static ESStoreProductViewControllerManager *_sharedManager = nil;
+    static dispatch_once_t onceToken;
+    dispatch_once(&onceToken, ^{
+        _sharedManager = [[self alloc] init];
+    });
+    return _sharedManager;
+}
 
 + (BOOL)storeProductViewControllerExists
 {